Fix warning (which is treated as an error) in trace_event

+++ base/debug/trace_event.h
@@ -720,8 +720,7 @@ class TraceID {
   };
 
   explicit TraceID(const void* id, unsigned char* flags)
-      : data_(static_cast<unsigned long long>(
-              reinterpret_cast<unsigned long>(id))) {
+      : data_(reinterpret_cast<unsigned long long>(id)) {
     *flags |= TRACE_EVENT_FLAG_MANGLE_ID;
   }
   explicit TraceID(ForceMangle id, unsigned char* flags) : data_(id.data()) {
